refactor: more UBID -> ZLDAP refactor pt.2
