--- store/src/main/java/com/zimbra/cs/ldap/unboundid/UBIDLdapClient.java
@@ -1,122 +1,0 @@
-// SPDX-FileCopyrightText: 2022 Synacor, Inc.
-// SPDX-FileCopyrightText: 2022 Zextras <https://www.zextras.com>
-//
-// SPDX-License-Identifier: GPL-2.0-only
-
-package com.zimbra.cs.ldap.unboundid;
-
-import com.zimbra.common.service.ServiceException;
-import com.zimbra.cs.ldap.LdapClient;
-import com.zimbra.cs.ldap.LdapConstants;
-import com.zimbra.cs.ldap.LdapException;
-import com.zimbra.cs.ldap.LdapServerConfig.ExternalLdapConfig;
-import com.zimbra.cs.ldap.LdapServerType;
-import com.zimbra.cs.ldap.LdapUsage;
-import com.zimbra.cs.ldap.ZLdapContext;
-import com.zimbra.cs.ldap.ZLdapFilterFactory;
-import com.zimbra.cs.ldap.ZMutableEntry;
-import java.util.Date;
-
-public class UBIDLdapClient extends LdapClient {
-
-	private final UBIDLdapPoolConfig poolConfig;
-
-	private UBIDLdapClient(UBIDLdapPoolConfig poolConfig) {
-		this.poolConfig = poolConfig;
-	}
-
-	public static UBIDLdapClient createNew(UBIDLdapPoolConfig poolConfig) throws LdapException {
-		ZLdapFilterFactory.initialize();
-		final ZLdapFilterFactory ubidLdapFilterFactory = new ZLdapFilterFactory();
-		ZLdapFilterFactory.setInstance(ubidLdapFilterFactory);
-		return new UBIDLdapClient(poolConfig);
-	}
-
-	public static UBIDLdapClient createNew(boolean alwaysUseMaster) throws LdapException {
-		final UBIDLdapPoolConfig ldapPoolConfig = UBIDLdapPoolConfig.createNewPool(alwaysUseMaster);
-		return createNew(ldapPoolConfig);
-	}
-
-	@Override
-	protected void terminate() {
-		poolConfig.shutdown();
-	}
-
-	@Override
-	protected void forceUsingMaster() {
-		poolConfig.setReplicaToMasterPool();
-	}
-
-	@Override
-	protected ZLdapFilterFactory getLdapFilterFactoryInstance()
-			throws LdapException {
-		ZLdapFilterFactory.initialize();
-		return new ZLdapFilterFactory();
-	}
-
-	@Override
-	protected void waitForLdapServerImpl() {
-		while (true) {
-			ZLdapContext zlc = null;
-			try {
-				zlc = new ZLdapContext(poolConfig, LdapServerType.REPLICA, LdapUsage.PING);
-				break;
-			} catch (ServiceException e) {
-				// may called at server startup when logging is not up yet.
-				System.err.println(new Date() + ": error communicating with LDAP (will retry)");
-				e.printStackTrace();
-				try {
-					Thread.sleep(LdapConstants.CHECK_LDAP_SLEEP_MILLIS);
-				} catch (InterruptedException ie) {
-				}
-			} finally {
-				if (zlc != null) {
-					zlc.closeContext(false);
-				}
-			}
-		}
-	}
-
-	@Override
-	protected ZLdapContext getContextImpl(LdapServerType serverType, LdapUsage usage)
-			throws ServiceException {
-		return new ZLdapContext(poolConfig, serverType, usage);
-	}
-
-	/**
-	 * useConnPool is always ignored
-	 */
-	@Override
-	protected ZLdapContext getContextImpl(LdapServerType serverType, boolean useConnPool,
-			LdapUsage usage)
-			throws ServiceException {
-		return getContextImpl(serverType, usage);
-	}
-
-	@Override
-	protected ZLdapContext getExternalContextImpl(ExternalLdapConfig config, LdapUsage usage)
-			throws ServiceException {
-		return new ZLdapContext(config, usage);
-	}
-
-	@Override
-	protected ZMutableEntry createMutableEntryImpl() {
-		return new ZMutableEntry();
-	}
-
-	@Override
-	protected void externalLdapAuthenticateImpl(String[] urls,
-			boolean wantStartTLS, String bindDN, String password, String note)
-			throws ServiceException {
-		ZLdapContext.externalLdapAuthenticate(urls, wantStartTLS,
-				bindDN, password, note);
-	}
-
-	@Override
-	protected void zimbraLdapAuthenticateImpl(String bindDN, String password)
-			throws ServiceException {
-		// TODO: check me
-		ZLdapContext.zimbraLdapAuthenticate(bindDN, password, poolConfig);
-	}
-
-}
